backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.12
firebase-admin==6.3.0
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.32.3
